+++ Sources/SwiftUIKit/_Deprecated/Images/Image+Deprecated.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+public extension Image {
+
+    @available(*, deprecated, renamed: "init(image:)")
+    init(imageResource: ImageRepresentable) {
+        #if os(iOS) || os(watchOS) || os(tvOS)
+        self.init(uiImage: imageResource)
+        #elseif os(macOS)
+        self.init(nsImage: imageResource)
+        #endif
+    }
+}