+++ Sources/SwiftUIKit/_Deprecated/Images/ImageResource.swift
@@ -1,0 +1,2 @@
+@available(*, deprecated, renamed: "ImageRepresentable")
+public typealias ImageResource = ImageRepresentable