+++ Release Notes.md
@@ -11,13 +11,18 @@ This will make the things in this library easier to understand and document, sin
 
 ### ✨ New features
 
+* `ColorRepresentable` is a new typealias that bridges NSColor and UIColor.
 * `ErrorAlerter` is a new protocol that can be implemented by any type that can alert errors using an `AlertContext`.
 * `ErrorAlertConvertible` is a new protocol that can be implemented by any error that can describe itself as an error `Alert`.
-
+* `FontRepresentable` is a new typealias that bridges NSFont and UIFont.
 
 ### 💡 Behavior changes
 
 * `PresentationContext` is now `open` instead of `public`.
+
+### 🗑 Deprecations
+
+* `ImageResource` has been renamed to `ImageRepresentable`.